Removing enable_utf8 config setting.
The reason it was added in r1250 was to make Kohana load faster, which is a valid reason. However, UTF-8 support is integrated throughout Kohana and should always be supported. To achieve that while cutting out a lot of load time, the utf8 functions now moved to separate files that are only loaded when needed.

My benchmarks on the loading of utf8.php went from about 0.0072 to 0.0030 seconds.